+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bundled compose file now utilises [docker-socket-proxy](https://github.com/linuxserver/docker-socket-proxy)
   for enhanced security
 - Automatically negotiate API version with the connected docker daemon
+- Blacklist container health reporting for a container by applying the `docker-prometheus-exporter.metric.container_health.enabled=false` 
+  label
 
 ## [1.0.0] - 2025-04-15